requirements.txt
python-telegram-bot
jdatetime
aiosqlite
pytz
asyncping3